v0.2.0: Integer overflow protection + wider API
Critical fixes:
- e12_norm returns int64_t (was int32_t — overflows for |a| > 46340)
- e12_mul uses int64_t intermediates (was int32_t — UB on overflow)
- New e12_mul_checked: returns false on overflow
- New e12_widen + e12_norm_wide for explicit 64-bit path
- Overflow check helpers: e12_would_overflow_add, e12_would_overflow_mul
- Added <stdbool.h> include

All 18 tests still passing with -Wall -Wextra -Werror -pedantic

+++ test_eisenstein.c
@@ -48,7 +48,7 @@ int main(void) {
 
     /* Rotation preserves norm */
     r = a;
-    int32_t n0 = e12_norm(a);
+    int64_t n0 = e12_norm(a);
     for (int i = 0; i < 100; i++) {
         r = e12_rotate_60(r);
         if (e12_norm(r) != n0) break;
